TutorialRecommender/requirements.txt
crewai>=0.1.32
langchain>=0.1.0
python-decouple>=3.8.0
python-dotenv>=1.0.0
requests>=2.31.0
